refactor(providers): let each provider own its alerts
alerts.ts knew about every provider payload, so touching one provider meant editing a shared
file. WeatherProvider now declares getAlerts(), empty by default, and each provider overrides
it with its own request and mapping — Open-Meteo simply inherits the default.

Météo France and OpenWeather read their alerts from a payload getWeather already fetched, so
they take it as an optional third argument: calling getAlerts() on its own still works without
requesting it twice.

The provider mappings move next to their provider (owmAlerts.ts, accuAlerts.ts, and the
vigilance severity helpers into mfWarnings.ts) along with their tests; alerts.ts keeps only
what is provider-agnostic.
