+++ app/services/providers/weatherprovider.ts
@@ -1,7 +1,7 @@
 import { WeatherLocation } from '../api';
 import { WeatherProps } from '../weatherData';
 import { Provider } from './provider';
-import { AirQualityData, WeatherData } from './weather';
+import { AirQualityData, Alert, WeatherData } from './weather';
 
 export interface GetWeatherOptions {
     model?: string;
@@ -14,6 +14,16 @@ export interface GetWeatherOptions {
 }
 export abstract class WeatherProvider extends Provider {
     abstract getWeather(weatherLocation: WeatherLocation, options?: GetWeatherOptions): Promise<WeatherData>;
+    /**
+     * Weather alerts for that location, called by `getWeather`. A provider owns both the request
+     * and the mapping of its own payload; the ones without an alert API keep this default.
+     *
+     * A provider whose alerts come from a payload `getWeather` already fetched takes it as a third
+     * argument, so calling `getAlerts` on its own stays possible without requesting it twice.
+     */
+    getAlerts(weatherLocation: WeatherLocation, options?: GetWeatherOptions): Promise<Alert[]> {
+        return Promise.resolve([]);
+    }
     getModels() {
         return {};
     }
